+++ app/Resources/TwigBundle/views/Exception/error404.html.twig
@@ -8,6 +8,8 @@
             <p>Le web a aussi ses tricks</p>
             <p>celui là s'appelle un <strong>"404" !</strong></p>
         </div>
-        <a href="{{ path('homepage') }}">Retourner à l'accueil</a>
+        <div class="btn text-center">
+            <a href="{{ path('homepage') }}">Retourner à l'accueil</a>
+        </div>
     </div>
 {% endblock %}
