+++ src/AppBundle/Controller/CommentController.php
@@ -4,7 +4,7 @@ namespace AppBundle\Controller;
 
 use AppBundle\Entity\Trick;
 use AppBundle\Entity\Comment;
-use AppBundle\Form\CommentType;
+use AppBundle\Form\Type\CommentType;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
