+++ src/AppBundle/Form/Type/PhotoType.php
@@ -1,31 +1,31 @@
 <?php
 
-namespace AppBundle\Form;
+namespace AppBundle\Form\Type;
 
-use AppBundle\Entity\Video;
+use AppBundle\Entity\Photo;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\UrlType;
-use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\Extension\Core\Type\FileType;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Form\FormBuilderInterface;
 
-class VideoType extends AbstractType
+class PhotoType extends AbstractType
 {
-    /**
-     * {@inheritdoc}
-     */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $builder->add('url', UrlType::class, array(
-        ));
+        $builder
+            ->add('file', FileType::class, array(
+                'label'=> false
+            ))
+        ;
     }
-    
+
     /**
      * {@inheritdoc}
      */
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults(array(
-            'data_class' => Video::class
+            'data_class' => Photo::class
         ));
     }
 
@@ -34,6 +34,8 @@ class VideoType extends AbstractType
      */
     public function getBlockPrefix()
     {
-        return 'appbundle_video';
+        return 'appbundle_photo';
     }
+
+
 }
