+++ src/AppBundle/Form/Type/ProfileType.php
@@ -1,7 +1,7 @@
 <?php
 // src/AppBundle/Form/ProfileType.php
 
-namespace AppBundle\Form;
+namespace AppBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
