+++ web/css/style.css
@@ -507,6 +507,30 @@ p.para{
 	-o-box-shadow: 0px 0px 5px #DDDDDD;
 }
 /***** Media Queries *****/
+@media only screen and (max-width: 1366px) {
+	.wrap{
+		width:95%;
+	}
+	.chute{
+		bottom: 30%;
+		right: 20%;
+	}
+	.chute h1{
+		font-size: 1.5em;
+	}
+}
+@media only screen and (max-width: 1280px) {
+	.wrap{
+		width:95%;
+	}
+	.chute{
+		bottom: 30%;
+		right: 5%;
+	}
+	.chute h1{
+		font-size: 1.5em;
+	}
+}
 @media only screen and (max-width: 1024px) {
 	.main_bg{
 		padding: 50px 100px;
